moved mutation code into seperate class, split config object into files
