+++ src/layout_config/default_state_mappings.js
@@ -1,0 +1,10 @@
+export default {
+  'ACT': 'Australian Capital Territory',
+  'NSW': 'New South Wales',
+  'NT' : 'Northern Territory',
+  'QLD': 'Queensland',
+  'SA' : 'South Australia',
+  'TAS': 'Tasmania',
+  'VIC': 'Victoria',
+  'WA' : 'Western Australia'
+}
