+++ src/layout_config/optimized_one_page_checkout.js
@@ -1,0 +1,37 @@
+export default {
+  label: "Optimized one-page checkout (Early access)",
+  layoutSelector: "#micro-app-ng-checkout",
+  countryIdentifier: 'countryCodeInput',
+  searchIdentifier: "addressLine1Input",
+  nz: {
+    countryValue: "string:NZ",
+    elements: {
+      address1: 'addressLine1Input',
+      suburb: 'addressLine2Input',
+      city: 'cityInput',
+      region: 'provinceInput',
+      postcode: 'postCodeInput'
+    },
+    regionMappings: null
+  },
+  au: {
+    countryValue: "string:AU",
+    elements: {
+      address1: 'addressLine1Input',
+      address2: 'addressLine2Input',
+      suburb: 'cityInput',
+      state: 'provinceCodeInput',
+      postcode: 'postCodeInput'
+    },
+    stateMappings: {
+      'ACT': 'string:ACT',
+      'NSW': 'string:NSW',
+      'NT' : 'string:NT',
+      'QLD': 'string:QLD',
+      'SA' : 'string:SA',
+      'TAS': 'string:TAS',
+      'VIC': 'string:VIC',
+      'WA' : 'string:WA'
+    }
+  }
+}
